requirements.txt
tweepy==4.14.0
python-linkedin-v2==0.7.11
mastodon.py==1.8.1
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3


